+++ lib/src/channels/channels.dart
@@ -5,3 +5,4 @@ export 'channel_client.dart';
 export 'channel_id.dart';
 export 'channel_video.dart';
 export 'username.dart';
+export 'video_sorting.dart';